fix: prevent mobile interactions from triggering the virtual keyboard
